TXLib: Turn on -Wall and fix all wanings

+++ modules/TXResource/src/io_file.cpp
@@ -2,7 +2,9 @@
 // Module: TXResource
 
 #include "impl/io_file.hpp"
+#include <cstddef>
 #include <fstream>
+#include <filesystem>
 
 namespace tx {
 
@@ -14,7 +16,7 @@ BinaryArray readWholeFileBin(const std::filesystem::path& filePath) {
 	std::ifstream ifs(filePath, std::ios::binary);
 	if (!ifs) { throw std::runtime_error("tx::readWholeFileBin(): Failed to open file"); }
 	ifs.read(reinterpret_cast<char*>(buffer.data()), fileSize);
-	if (ifs.gcount() != fileSize) { throw std::runtime_error("tx::readWholeFileBin(): Failed to read entire file. Buffer size not matching file size."); }
+	if (static_cast<size_t>(ifs.gcount()) != fileSize) { throw std::runtime_error("tx::readWholeFileBin(): Failed to read entire file. Buffer size not matching file size."); }
 	return buffer;
 }
 std::string readWholeFileText(const std::filesystem::path& filePath) {
@@ -25,7 +27,7 @@ std::string readWholeFileText(const std::filesystem::path& filePath) {
 	std::ifstream ifs(filePath, std::ios::binary);
 	if (!ifs) { throw std::runtime_error("tx::readWholeFileText(): Failed to open file"); }
 	ifs.read(buffer.data(), fileSize);
-	if (ifs.gcount() != fileSize) { throw std::runtime_error("tx::readWholeFileText(): Failed to read entire file. Buffer size not matching file size."); }
+	if (static_cast<size_t>(ifs.gcount()) != fileSize) { throw std::runtime_error("tx::readWholeFileText(): Failed to read entire file. Buffer size not matching file size."); }
 	return buffer;
 }
 } // namespace tx
